Change: serde optional behind feature toggle
To allow multiple serializer the serde macro usage is behing a feature
toggle. This allows to use models without serde support and with that
the possibility to write own serializer and work with the same data
structure.

The feature toggle is `serde_support` and is enabled on default.

Additionally the json module is dropped as it is not limited to json.