Change: switiching from rocket to hyper
Instead of asking much is the fish we went hyper, hyper.

Replaces the rocket implementation to hyper so that we have more fine
granular control.

Since the scanner-api definition is rather simple there are not that
many endpoints to consider which makes it feasible to user hyper
directly.

To change the endpoints you need to change the entrypoint function
within entry.rs. There are two background tasks starting on
initialization:
- results/fetch
- feed/fetch

`results/fetch` does periodically check ospd-openvas for results of
started scans while `feed/fetch` verifies if the feed got updated and
fetches all oids of that feed for the `vts` endpoint.

Sensord utilizes `osp` crate to control ospd.

Co-authored-by: Christoph Krämer <c.kraemer.rulle@web.de>