Fix: CI: signature of tar balls, trigger release without tag
Fixes the signature of tar balls by using
`curl -sfSL --retry 3 --retry-connrefused --retry-delay 2`
when fetching the tar balls of the created release.

This changes the curl statement to
- `s` - siltently
- `f` - fail on server error
- `S` - showing error
- `L` - following redirects
- `retry 3` - try three times before giving up on failure
- `retry-connrefused` - retry on connection refused
- `retry-delay 2` - wait two seconds before a retry
this should make the curl statement more resilient.

It fixes the failing triggering on a merge of pull request when no
release label was provided by removing `${{ ... }}` in favor of `( ...
)` if clause in release description. Removing failed runs on `release`.