Fix: Docker: x86_64 nasl-cli
Changes the cross.Dockerfile to also build the dependencies as in the
arch64 versions so that we can control the libssh version.

It fixes a bug with zlib and uses zlib.so.1 instead of zlib.so to enable
usage of zlib1g instead of zlib1g-dev.

It also moved to building of the rust binaries to `build-rust.yml` which
is used in container, rustification as well as release. This ensures
that the binaries are build in the same way everywhere.