Change: rename queued to stored
Instead of queued we use stored to describe that a scan is stored but
not yet started.

This means a scan is stored it gets the status stored and when started
it switches to requested until the scan is actually running and then it
will be in the state running.

- stored => scan is stored
- requested =>  scan is started
- running => scan is running
- stopped | failed | succeeded => scan is done