Fix: error handling, when ospd socket was not found
In case the OSPD socket was not found, openvasd printed an error, but set the scan as requested and sent regular requests to OSPD to get its status. In case of an error, the scan does not change its status and not requests to OSPD are sent.

Small improvement of the logging. In case the OSPD socket is not available, it is logged on startup, with the remark, that some functions are not available.