Fix: override an existing scan only
This patch solves an issue in which a running scan could be overriden if a new scan is post/stored with the same scan ID, producing a reset of the running scan status and results lost
Instead, the API changed and now it is not possible to provide a scan id when creating a scan.
The scan id is automatically generated by openvasd and sent back to the client in the response.