Add: POC for integrating C based NASL libraries
This PR introduces a concept for implementing c based nasl libraries within rust.
This works by a build script, which uses the rust cc library to compile the c code and link it to the created rust library.

This POC also contains a solution integrating the GMAC algorithm, as there is currently no support for it in the RustCrypto project.