Add: openvasd: encryption of credentials and results
To allow easier handling of sensitive data the direct usage of a hash
map is replaced with a storage interface and field encryption is added.

This is done by introducing interfaces:
- to get the progress of a scan
- store scans
- append results

instead of using a hashmap of a progress struct. This allows us to
exchange the storage from inmemory to something like a relational
database.

Additionally a crypt interface with a chacha20 implementation is
introduces to enable encryption of sensitive fields like passwords or
results.