Add: infisto as a file storage solution
To be able to store data on disk to be more resilient on daemon errors a
new library `infisto` (indexed file storage) is created and used in
openvasd.

This allows with the configuration of openvasd

```
[storage]
type = "fs"

[storage.fs]
path = "/tmp/openvasd/storage"
key = "see config.example.toml for more details"
```

to store scans and their results on encrypted on disk.

Although the key used for encryption can should set via the environment
variable `STORAGE_KEY` of the system user that runs openvasd rather than
in the configuration or start parameter each way is supported.