Change: overhaul Statement
Redefines Statement from an enum containing all information to a struct.

This Statement struct contains a start and end token as well as a
`StatementKind`. The kind contains all necessary information for
execution.

The Statement also contains information about start and end to make it
also useful when working with code manipulation e.g. transpiling.

This refactoring allows us to extend Statement information, for e.g. a
formatter, while not having to change the interpreter itself.

Overall this should give us more freedom to extend Statements.