Change: move verification into Iteration Item
For the verification of the Hashsum, used in the feed, an Iterator is used to check each line of the Hashsum file. Before the Hashsum was immediately verified, regardless if each iteration Item was even necessary. With this change, it is possible to only verify the Files, when they are needed.