Change: renaming of advisories to product and advisory to vt
In order to reduce confusing naming schemes the name advisory is completely dropt for the notus implementation and replaced by either product, when
referencing the file or collection of advisories and to vulnerability tests when tailking about the actual tests that are executed during a scan.

Added missing developer documentation for functions, structs, etc.

Added Readme for the notus rust library