Change: toggle smoketest
Instead of having to run a smoke test for each cargo test we just run it
when explicitely set. This is done by introducing a feature called
`smoketest` that is toggling if the `smoketest/tests/tests.rs` is run.

To run the smoke tests execute:

```
cargo test --features=smoketest
```

this is also reflected within the Makefile.