Fix: helm: notus-data uses a tar.gz file
Within the latest changes of the notus-feed it was decided that an image
contains just the tar.gz file, forcing us to utilize the init.sh script.

Since the same is planned for vulnerability-tests the deployment is also
changed to utilize the init.sh script instead of doing that as a
command.