Change: update hyper to version 1
This changes the underlying version of hyper to version 1.

The run methods is moved to controller. This hypothetically allows us to
run a complete server within tests.

When TLS is enabled, which should be the case for all production
environments, then it will use http2. On http it uses http1 so that it
is more convenient to use with curl.

The stream handling of results and oids is changed to channel and chunks
as it is easier to understand and requires less code then our previous
Stream approach.

The handling of TLS and insecure mode are now more separated, although
this results in wet code with a few duplication it is believed to be
easier to change when there is a problem within one mode but not the
other.

The client identifier handling is also changed, rather then using
wrapper upon wrapper we use a global arc in run and reset per request.