Change: set Scan ID of Scan Model to non-optional
As the scan ID was set by the server and not by the client, it was set to Optional.
This lead to many unnecessary checks, even though it was not possible, that this value is None.
The new implementation automatically fills in an UUID for the scan ID, when parsing a Scan Config.