Add: scan scheduler
Changes context to contain only a scheduler and not any scanner or
storage directly. This to ensure that each command goes through the
scheduler so that it can keep track if a feed update is running or a
scan is queued.

Additionally the background processes call the scheduler to synchronize
scans and feed. A feed can only be updated when no scans are running and
a scan only started when no feed update is running.

This means essentially that the Scheduler within scheduling plays a core
role in openvasd shifting away from direct scanner access.

Since a scheduler uses only standard interfaces of models::scanner each
scanner needs to keep track by themselves depending on their technical
solution. In a nutshell this means that each scanner must be implemented
within it's crate:
- osp has a scanner implementation
- openvasctl has a scanner implementation