Change: remove FetchResult in favor of ScanResults
Instead of using a tuple of (Status, [Result]) a struct ScanResults is
introduced. This will help us to create bulk retrieving operations on
implementations that makes sense.

Although the benefits are currently minimal when we introduce a
scheduler for scans it will make it easier to handle results in that
fashion.