Fix: multiple verbose entries in nasl-cli
Verbose is set on the overall clap command although it should be set on
the subcommand itself. This leads to
```
debug_asserts.rs:86:13:
Command nasl-cli: Argument names must be unique, but 'verbose' is in use by more than one argument or group
```
when it is not build with --release.

This commit changes this behaviour by calling it on the newly build
subcommand rather than the given root command.