CI: release cannot be dependent on smoketest
Currently we don't push images based on a pr but just based on main or
an version tag. This means that on a closed PR with a release there is
no new docker image pushed and the smoketest is not executed.

To mitigate it the release is not dependent on a smoketest but on
container build. Alternatively we should either make the smoketest not
dependent on a container image push or push the image when creating the
release instead of when we create the github tag.