Add: nasl-interpreter scheduling
Adds possibility to schedule script execution based on a Scan configuration.
This scheduling will load all script_dependencies and sort them to stages.
Each list of scripts called by `next()` can be executed concurrently.
The iterator implementation should make it hard to implement a solution that
tries to execute scripts that are not allowed to run concurrently.

It is implemented as a Extension for each storage::Retriever<K> where K has a
Default implementation.

Currently it only implements the WaveExecutionPlan.
It is based on script_dependencies if a script does not have any dependencies
then it is add into index 0, if it has dependencies that are mentioned in
index 0 then it will be inserted into index 1, if it has dependencies that are
in index 1 then it will be inserted into index 2 and so on. It does not care
for individual dependencies and just orders it into a wave.

Example:
 ```

 use nasl_interpreter::scheduling::ExecutionPlaner;
 use nasl_interpreter::scheduling::Stage;
 use nasl_interpreter::scheduling::WaveExecutionPlan;
 use storage::Dispatcher;
 use storage::Retriever;

 let feed = vec![
     storage::item::Nvt {
         oid: "0".to_string(),
         filename: "/0".to_string(),
         ..Default::default()
     },
     storage::item::Nvt {
         oid: "1".to_string(),
         filename: "/1".to_string(),
         dependencies: vec!["/0".to_string()],
         ..Default::default()
     },
     storage::item::Nvt {
         oid: "2".to_string(),
         filename: "/2".to_string(),
         dependencies: vec!["/1".to_string()],
         ..Default::default()
     },
 ];
 let retrieve = storage::DefaultDispatcher::new(true);
 feed.clone().into_iter().for_each(|x| {
     retrieve
         .dispatch(&String::new(), x.into())
         .expect("should store");
 });

 let scan = models::Scan {
     vts: vec![models::VT {
         oid: "2".to_string(),
         parameters: vec![],
     }],
     ..Default::default()
 };
 let results = (&retrieve as &dyn Retriever<String>)
     .execution_plan::<WaveExecutionPlan>(&scan)
     .expect("no error expected");
 assert_eq!(
     vec![
         (Stage::End, vec![(feed[0].clone(), None)]),
         (Stage::End, vec![(feed[1].clone(), None)]),
         (Stage::End, vec![(feed[2].clone(), Some(vec![]))]),
     ],
     results.filter_map(|x| x.ok()).collect::<Vec<_>>()
 )
 ```