WIP: simplify storage
The previous storage system was thinking in terms of streaming fields to make
it easier to adapt storage forms within our implementation.

The problem with this solution is however that it makes it very confusing when
working with static non field but item based elements that you mostly want to
consume as a whole rather then per field basis (e.g. VT meta information).

As a compromise the concept of ItemDispatcher was born. Meant as a work around
for VTG specifically through incorrect naming and ease of use it got adapted in
redis implementation and openvasd. However the scannerctl was using the field
based storage style.

That means that a developer now has to choose which model to follow, making it
opaque and frustating.

This refactoring effetively removes the field bases storage system and just
allows it for VT (because it is necessary) by combining it into one trait that
each storage must implement in order to be used instead of three traits.

Although the downside is that with each new storage type we need to adapt each
storage implementation it is much easier to use and more explicit when
developing as you don't have to choose anymore.