Refactor scheduling/mod to collect all dependencies before hand
To not have to cache each dependency within the actual implementation
scheduler does collecting a hashmap of string/vt and borrows it to the
scheduling implementation. That way the scheduler can  immediately
lookup if depedencies are missing instead of having to wait until next
is called and it doesn't have to create a shadow lookup table itself.