Fix: Option.unwrap on None value when unknown preference is set
As the preference handling does not return an Result but is written in a
way that it should never fail we should not assume that the customer has
set all preferences correctly but must ignore invalid settings instead.

This changes the first assumption and validates if the preference is
available and if not ignores it.