Add: scheduling scan to scannerctl
Usage: `scannerctl execute scan [OPTIONS] --path <FILE> [json]`

This is in a very early state. We need to add storage options for ports
or the whole scan to look them up when running.

We additionally need to add a alive checks that will set the required
parameter.

The script_dependency check is also not enabled for now.