Refactor: remove generic K in favor of enum
To make it clearer when to use what a ContextKey to define the scope of
the Retrieve, Dispatcher is introduced. ContextKey is a enum containing
either FileName or Scan with a String value for either a Scan context or
a feed update context.

The reason to introduce a enum over a String value is to make the
context more clear and to allow key verifications later on.