Refactor: Context owns the key, feed#Update uses references
To make feed#Update easier to use it does not own the given storage
nor the file loader anymore but borrows it. This makes it easier to use
feed#Update with shared contextx.

ContextBuilder is renamed to ContextFactory and does not store target
nor scan_id anymore. As each context is build for a new scan with a new
target Context now owns the scan_id as well as the target and both
parameter got removed from ContextFactory. This makes it possible to
share the factory while creating bound contexts.