Add: storage implementatoin for naslinterpreter#scanner
Since we need to share within openvasd storages by Arc and usually the
storage implementations are not written as mut we should not need a
lock.

Therefore Arc implementations for storage::Retrieve, store::Dispatcher,
openvasd#storage#Storage are implemented. This allows to create and
store and Arc storage.

Additionally todo implementation of storage::Storage are done for:
- file#Storage
- redis#Storage
within openvasd.

InMemory storage of openvasd has an underlying nasl-storage that is used
for everything but Results.

To prevent trying to start a new thread while async thread is running to
block the current thread which would be required to implement the
storage::Storage traits the tokio::sync::RwLock are replaced with
Arc<std::sync::RwLock> instead.

Unfortunately this means that each lock opration should be run within a
tokio::spawn blocking to not accidentally block the main thread of
tokio.