Fix: host handling within scan_interpreter
Because the scan_interpreter previously removed a VT when it was
executed,  it essentially only processed one host as tehre we no VTs
left for any host after that.

This is fixed by caching all VTs and keep track of:

- stage index
- VT index
- host index

that way, after the first host is done, we have each Stage step stored
in a Vector that in return contains all VTs that we then can lookup
based on the indices.