Add: storage.scan_finish and storage.clean_scan
To allow the nasl-interprete scanner to delete kb items as well as
results as new trait `Remover` is added and an function abstraction
within `Storage` are added. This allows the scanner implementation to
allow kb items when a scan is finished as well as the customer to delete
results based on that implementation.