refactor Sessions to yield borrowed sessions.
This complication is necessary since we acquire the MutexGuard for the
whole Vec but ideally want to return only a MutexGuard for the relevant
session. This isn't possible but in order to still have the convenience
of clean external function calls, we create a BorrowedSession guard
which contains both the MutexGuard for the Vec as well as the index of
the session within the Vec. This concrete implementation is done in
order to avoid problems with self-referential structs that we would
otherwise run into.