Fix: add target to kb limitation while not add target for results
After each host nasl-interpreter/scanner does delete the KB items to be
able to run various target concurrently the ContextKey::Scan is added
with an optional target. The DefaultDispatcher stores KBs with the whole
Contextkey so that they are target and scan specific while  results are
stored with ContextKey::value that just returns the ScanID.

This fixes the bug that no results did show up when a scan was finished
the results got stored as `ScanID-Target` and it does get rid of the
Scan not found bug within scan_interpreter based on the same underlying
issue.