Make nasl_function procmacro generate async fns.
Before, the macro generated an inner closure that captures
all the outer arguments and returns the result. Since async
closures are unstable, we have to slightly shift approaches
and generate an actual function (its name is generated by
modifying the name of the actual nasl_function) with the proper
signature. Depending on the receiver type and asyncness of the nasl
function, different function signatures are generated. This function
is then called instead of the closure.