Allow functions with &mut access to state.
We can differentiate between stateful functions that take `&State` as an
argument and those that take `&mut State`. By wrapping the state in the
`StoredFunctionSet` in an RwLock, we can access it via read or write,
depending on the function type.

1. This significantly improves ergonomics for implementing NASL functions,
since the state does not need to manage its own interior mutability
anymore.

2. This allows things that weren't previously possible. NASL functions
could only take &State arguments. A function managing a list of
connections `Vec<Connection>`, for example, would have to wrap the entire
list in an `RwLock` (`RwLock<Vec<Connection>>`) in order to allow adding and
deleting items of the list. With the new system, the state can simply be
`Vec<Connection>` if we never need mutable access to the connection, or
alternatively `Vec<RwLock<Connection>>` if we do need mutable access to the
connection.
This is an important step to allow true concurrency for these types of
functions, since previously every single function call would block the
entire state, while the new version only locks the entire state if a
function call requires mutable state.