Add: use `IpAddr` as type for the target of the context
As many NASL functions use the IP of the currently scanned target, it is a huge benefit to have the target already as an `IpAddr`,
as it had to be parsed every time used. This also removes the Optional target String, as the default target should always be
`127.0.0.1`. Also for scans, there always must be a target IP for scanning.