Fix context and register args with positionals.
Previously, the context and the register increased the argument count
and would increase the index of any subsequent positional arguments.
Now, they are not counted, as it should be. This also adds an error
message if the Context or the Register are not in front of the
positional arguments and imposes strict requirements for the ordering of
all argument types, namely
1. Context/Register
2. Positionals
3. Maybe positionals (i.e. arguments that can also be named)
4. Named arguments
5. Argument lists i.e. `Named` and `Positionals`.