Improve ergonomics for `TestBuilder`.
Improve ergonomics for `TestBuilder` by immediately checking the
reference result against a result as it comes in, not only once we
have obtained all the results. This means that if an intermediate
check already fails, we won't be performing any of the following lines
of code anymore. This helps finding the issue, since those lines might
otherwise deadlock or take forever if we're dealing with some kind of
async library.