Wrap the libssh `Channel` in a mutex.
This is necessary in order to unify the structure between the libssh and
the russh implementation. We want the `nasl_function` signature to be
the exact same, but this requires making the function on `SshSession`
have a consistent signature between the two implementations. Since the
function is necessarily `async` in `russh`, this means it also has to be
async in `libssh`. However, making any function on the libssh `SshSession`
async will automatically make it borrow `&self`. However, if the `channel`
contained in `SshSession` is not contained in a `Mutex`, `SshSession`
will be `!Send`. Thus, in order to have a clean interface, we need to
wrap the `Channel` in a `Mutex`.