Fix: nasl builtin function recv timeout after first response
recv may return before length bytes have been read: as soon as at least one byte
has been received, the timeout is lowered to 1 second. If no data is received
during that time, the function returns the already read data; otherwise, if the
full initial timeout has not been reached, a 1 second timeout is re-armed and
the script tries to receive more data from the socket. This special feature was
implemented to get a good compromise between reliability and speed when
openvas-scanner talks to unknown or complex protocols. This functionality was missing.