Fix: krb5_gss_update_context and okrb5_gss_free_context
In krb5_gss_update_context the data len was retrived by using strlen,
this is incorrect as the data is not a str, although the name indicates
it, but is actually a pointer to raw data and therefore
`get_var_size_by_num` must be used instead.

The freeing method `okrb5_gss_free_context` did not release gss specific
data this is changed.