Fix tests not running due to deadlock issues.
This is done by spawning a separate thread. The real problem is that
the libssh implementation pretends to be async but actually blocks on
every call. This is a bad situation (for concurrent performance) but as
long as the libssh calls finish, we won't deadlock due to it in normal usage
of openvasd. However, in the SSH tests, concurrency was used to simultaneously
run the client and the server, which led to problems if the client blocked on
a connect call (but never yielded to the runtime so that the server could continue).
This commit fixes this situation by running the server in a separate thread
explicitly.