Avoid dropping the TestBuilder in async_verify.
If the `verify` call in `async_verify` panicked, the the  `Drop` impl
would be called, which would itself panic (because we were in an async
context), obscuring the actual error message. A nice solution would
have been to use `catch_unwind`, but this is not allowed for async fns.
Instead, we make `verify` return an actual result and defer the panic.