Fix: `get_version()` of packages containing the epoch resulted in a leading `0:`
The leading `0:` is confusing when in the compared version this is not contained, even though it means the same thing, as a missing epoch is the same as `0`.

Additionally this contains a small refractoring of the package modules of notus, which removes duplicated data within the package struct. `full_version`
contained the information of different version parts in a single string. This is now generated on demand from the available information with `get_version()`.