Add: nasl-function: ip_reverse_lookup
To be able to reverse lookup the IP address of either the target or
given IP address a new function `ip_reverse_lookup` is introduced.

This function does not rely on found vhosts to be easier to use within a
krb5 context. It does not fork and just calls `getnameinfo`.

The function can used with an IP address:

```
ip_reverse_lookup('1.1.1.1');
```

and without:
```
ip_reverse_lookup();
```

When no ip address is given than the target ip address will be used.