Add `run_find_service` implementation.
This is a very rudimentary implementation of the `find_service` builtin. It only includes a limited number of services but hopefully lays the foundation of something ... great.

In this implementation, service detection is not hardcoded, but determined by a list of services in a JSON file. In this JSON, the services are defined as metadata (such as name and description), possible ports, along with some detection methods as well as custom KB entries to add when the service is detected.

For now, only a smaller number of services are detected:
FTP, telnet, SSH, web/http, SMTP, POP3, IMAP and NetBus.

For testing, a dockerfile along with a NASL script and a bash script to
start the dockerfile and run the NASL script against the container is
provided.