Add: report_scripts openvas setting
Stores in a file the script run duration in json format.
This option expects a valid path to store a file with scripts stats.
Script run durations are collected only if the log_whole_attack setting
is set to 'yes'.

For testing, set the setting in openvas.conf with a valid path and set log_whole_attack setting to yes.
A json file will be created containing script run duration for each host as json objects.