Add: Port handling within KB and acording NASL functions
This PR contains an extension of the retriever trait to also be able to get kb items with a pattern.
This is needed in order to get a list of all open ports of a target, as they are saved within the kb.

Additionally added the following NASL builtin functions:
- get_host_open_port
- get_port_transport

When creating an socket, now the acording transport used is entered into the kb