Removing impacket in oldstable for now
It breaks the oldstable build when creating an image for arm64.

Since it should not be used in production anyway and we just use it for
testing if we can still compile I think it is ok to remove runtime
depdencies for now at least until a solution is found.