CI: move building the rust binary inside docker image
To prevent issues when building arm64 and simplify the setup the building
of our rust code is moved from external source into the docker image.

This should make maintaing the images a bit easier in the cost of slower
image build time.

Additionally we get rid of the qemu setup in favor of a aarch64 based
internal runner.

Revert "Removing impacket in oldstable for now"

This reverts commit 61b911d004ee75bdbd8829b68a426171d673573b.