compile release on when necessary
Instead of building the binaries everytime it is build when they are
needed.

This speeds up the regular CI checks significantly in the costs of
slowing down functional tests and release.

As we release less often then building and usually just want to know if
the change compiles it is replaced with `cargo check`.