CI: remove unittests dependency for build
Removes unittests dependency on build to let the build and unit test run
concurrently.

Add unittests dependency for functional tests as it doesn't make sense
to run functional tests when the unit tests fail.

Removes functional dependency for container as it is not a requirement
for non release container to be functional.