Fix: isiplocalnet() nasl function. (#1842)
Jira: SC-1232
Close #1829
The function just check the first address in the network interface. This produces that if the target host is in the same network as a second address in a network interface returns a false negative.

E.g.: given the following config, if the target host has 192.168.10.100, the target was not recognize as in the same local network, because the local ip in the 192.168.10.0/24 is the second defined in the interface.
```
4: wlp6s0: <BROADCAST,MULTICAST,UP,LOWER_UP> mtu 1500 qdisc noqueue state UP group default qlen 1000
    link/ether cc:5e:f8:68:06:bd brd ff:ff:ff:ff:ff:ff
    inet 192.168.1.6/24 brd 192.168.1.255 scope global dynamic noprefixroute wlp6s0
       valid_lft 65706sec preferred_lft 65706sec
    inet 192.168.10.1/24 scope global wlp6s0
       valid_lft forever preferred_lft forever
    inet6 fe80::ec78:2c18:a29f:ef3a/64 scope link noprefixroute
       valid_lft forever preferred_lft forever
```

With this patch, all addresses in all interfaces are checked.