Prevent misuse of `context()` on `TestBuilder`.
The `context()` constructs a `Context` from the given data in the
`TestBuilder` and returns it. As a user of `TestBuilder`, it is tempting
to take the generated `Context` to set certain properties before running
the test. However, in the current structure, any changes made to the
context in this way will be discarded anyways, since the `Context`
will be rebuilt from scratch when the code is actually executed.

There is also no way to change the behavior (for example by returning a
`&mut Context`) in the current structure, because the `Context` has to
be created when `context` is called, so there is no lifetime we could
return, the only thing to do is to transfer ownership.

In order to prevent this misuse, this commit makes the `context` method
private. A test that uses it is rewritten to use a new
`results_and_context` method which returns the results of the code along
with the `Context` used to obtain them.