Fix: delete scan
When the client crashes and later tries to resume the scan with the same scan ID, two entries with the same scan id exists in the list. Trying to remove one, changes the length of the vector producing openvasd to crash. This is solved inverting the order of the elements to be remove, to ensure the element index to be removed is always less than the vector length

Jira: SC-1225