Use a scanner-internal representation of `Scan`.
This creates a separation between the outer representation
(`models::Scan`), which is used
1. for (De-)serialization
2. within osp/openvas/the openvasd API.

and the internal type used in the openvasd scanner
implementation.

This allows using different types and representations within the
internal representation. In particular, it allows to use the "real"
target type instead of a simple hostname-string, which in turn allows
moving the hostname resolution logic from an implicit resolution
somewhere deep in the call chain to an explicit declaration at each call
site of the scanner.

As a small bonus, the scanner unit tests are now 1000% faster ;)